requirements.txt
asgiref==3.7.2
coverage==7.3.1
Django==3.2.21
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
packaging==23.1
pip==23.3
pluggy==1.3.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==68.2.2
sqlparse==0.4.4
uritemplate==4.1.1
wheel==0.41.2
